Removed old model
Old data model has been removed, and has been replaced by the new one.

There's still bug, but it seems to be functional.
Memory mapped output works as intended.
Memory mapped input does not work (needs support from main window).
Speed has decreased (45.0MHz-4.6MHz).

All files in all classes now reference the abstract implementations, which should promote sharing between projects.

+++ pep9common/amemorydevice.h
@@ -15,12 +15,15 @@ public:
 
     bool hadError() const;
     QString getErrorMessage() const;
-    virtual quint16 size() const = 0;
+    virtual quint32 size() const = 0;
 
     void clearErrors();
 
     const QSet<quint16> getBytesWritten() const;
     const QSet<quint16> getBytesSet() const;
+    // Call after all components have (synchronously) had a chance
+    // to access these fields. The set of written / set bytes will
+    // continue to grow until explicitly reset.
     void clearBytesWritten();
     void clearBytesSet();
